feat(next): Renames next things nad PR fixes
