--- apps/next/src/app/pages/index-page/index-page.html
@@ -1,8 +1,0 @@
-<main>
-  <h1>NiceHeadline</h1>
-  <div class="ba-page-content" *ngIf="page" [innerHTML]="content"></div>
-  <div class="ba-page-content" *ngIf="!page">
-    <p>No content found</p>
-  </div>
-  <router-outlet></router-outlet>
-</main>
