+++ apps/next/src/app/pages/next-error-page/next-error-page.html
@@ -1,9 +1,9 @@
 <main>
   <h1>NiceHeadline</h1>
-  <div class="ba-page-content" *ngIf="content">
+  <div *ngIf="content">
     {{ content }}
   </div>
-  <div class="ba-page-content" *ngIf="!content">
+  <div *ngIf="!content">
     <p>No content found</p>
   </div>
 </main>
