+++ apps/next/src/app/pages/next-error-page/next-error-page.scss
