+++ apps/next/src/app/pages/next-index-page/next-index-page.html
@@ -1,0 +1,8 @@
+<main>
+  <h1>NiceHeadline</h1>
+  <div *ngIf="page" [innerHTML]="content"></div>
+  <div *ngIf="!page">
+    <p>No content found</p>
+  </div>
+  <router-outlet></router-outlet>
+</main>
