+++ apps/next/src/app/pages/next-index-page/next-index-page.scss
