+++ apps/next/src/app/pages/next-single-page/components/next-page-footer/page-footer.ts
