+++ apps/next/src/app/pages/next-single-page/next-single-page.html
@@ -7,7 +7,7 @@
 
   <main class="next-page-main" id="main" role="main">
     <article class="next-single-page-content">
-      <ba-page-content [data]="_pageContent.content"></ba-page-content>
+      <next-page-content [data]="_pageContent.content"></next-page-content>
     </article>
   </main>
