+++ apps/next/src/app/pages/next-single-page/next-single-page.scss
