+++ apps/next/src/app/styles/_page.scss
@@ -11,7 +11,7 @@
 }
 
 @media (min-width: $screen-md) {
-  .ba-page-grid {
+  .next-page-grid {
     grid-template-columns: minmax(0, 1fr) 230px;
     grid-column-gap: 2rem;
     grid-row-gap: 2rem;
@@ -22,13 +22,13 @@
     padding: 1rem;
   }
 
-  .ba-single-page-content::before {
+  .next-single-page-content::before {
     height: 2rem;
   }
 }
 
 @media (min-width: $screen-lg) {
-  .ba-page-grid {
+  .next-page-grid {
     grid-column-gap: 3rem;
     grid-template-areas:
       'title .'
@@ -39,7 +39,7 @@
 }
 
 @media (min-width: $screen-xxl) {
-  .ba-page-grid {
+  .next-page-grid {
     grid-template-columns: 300px minmax(0, 1fr) 280px;
     grid-template-rows: repeat(2, max-content) 1fr;
     grid-template-areas:
@@ -52,7 +52,7 @@
 
 /* All columns + column gaps (3rem) + padding-right (1rem) = 1738px */
 @media (min-width: 1740px) {
-  .ba-page-grid {
+  .next-page-grid {
     grid-template-columns: 300px auto 950px 280px auto;
     grid-template-areas:
       'sidenav . title . .'
