+++ apps/next/tslint.json
@@ -2,7 +2,7 @@
   "extends": "../../tslint.json",
   "rules": {
     "directive-selector": [true, "attribute", "x", "camelCase"],
-    "component-selector": [true, "element", "x", "kebab-case"],
+    "component-selector": [false, "element", "next", "kebab-case"],
     "dt-require-license-banner": true
   },
   "linterOptions": {
